nixos/syncthing: Trim the apiKeyFile because syncthing does not
If the file containing the API key also contained whitespace such as a
line break, jq would serialize this and syncthing would store the line
break and always return a 403 Forbidden.

Since this easily happens and syncthing does not warn about it (and you
can also not see the mistake in the text box on the web UI), we trim the
leading and trailing whitespace using jq's trim function (only available
from jq 1.8 & NixOS 25.11).